+++ R/np.mojo.R
@@ -2,8 +2,8 @@
 #' @description For a given lagged value of the time series, performs nonparametric change point detection of a possibly multivariate
 #' time series. If \code{lag} \eqn{\ell = 0}, then only marginal changes are detected.
 #' If \code{lag} \eqn{\ell \neq 0}, then changes in the pairwise distribution of \eqn{(X_t , X_{t+\ell})} are detected.
-#' @details The single-lag NP-MOJO algorithm for nonparametric change point detection is described in McGonigle, E. T. and Cho, H. (2023)
-#' Nonparametric data segmentation in multivariate time series via joint characteristic functions.  \emph{arXiv preprint arXiv:2305.07581}.
+#' @details The single-lag NP-MOJO algorithm for nonparametric change point detection is described in McGonigle, E. T. and Cho, H. (2025)
+#' Nonparametric data segmentation in multivariate time series via joint characteristic functions.  \emph{Biometrika}.
 #' @param x Input data (a \code{numeric} vector or an object of classes \code{ts} and \code{timeSeries},
 #' or a \code{numeric} matrix with rows representing observations and columns representing variables).
 #' @param G An integer value for the moving sum bandwidth;
@@ -78,7 +78,10 @@
 #'    \item{criterion, eta, epsilon}{Input parameters}
 #'    \item{test.stat}{A vector containing the NP-MOJO detector statistics computed from the input data}
 #'    \item{cpts}{A vector containing the estimated change point locations}
-#'    \item{p.vals}{The corresponding p values of the change points, if the bootstrap method was used}
+#'    \item{scores}{The corresponding importance scores of the estimated change points.
+#'    The larger the score is, the more likely that there exists a change point close to the estimated location.
+#'    If the bootstrap method is used, this a value between 0 and 1 corresponding to the proportion of times the observed detector statistic was larger than the bootstrapped detector statistics.
+#'    Otherwise, the importance score is simply the value of the detector statistic at the estimated change point location (which is not necessarily less than 1).}
 #' @references McGonigle, E.T., Cho, H. (2025). Nonparametric data segmentation in multivariate time series via joint characteristic functions. \emph{Biometrika}.
 #' @references Fan, Y., de Micheaux, P.L., Penev, S. and Salopek, D. (2017). Multivariate nonparametric test of independence. \emph{Journal of Multivariate Analysis},
 #' 153, pp.189-210.
@@ -93,7 +96,7 @@
 #' x <- signal + noise
 #' x.c <- np.mojo(x, G = 83, lag = 0)
 #' x.c$cpts
-#' x.c$p.vals
+#' x.c$scores
 #' @importFrom Rcpp evalCpp
 #' @useDynLib CptNonPar, .registration = TRUE
 #' @importFrom foreach %dopar%
@@ -194,7 +197,7 @@ np.mojo <- function(x, G, lag = 0, kernel.f = c("quad.exp", "gauss", "euclidean"
   }
 
   cpt.locs <- numeric(0)
-  p.vals <- numeric(0)
+  scores <- numeric(0)
 
   if (max(test.stat) > threshold.val) {
     exceedings <- (test.stat > threshold.val)
@@ -243,15 +246,15 @@ np.mojo <- function(x, G, lag = 0, kernel.f = c("quad.exp", "gauss", "euclidean"
     }
 
     if (threshold == "bootstrap") {
-      p.vals <- numeric(0)
+      scores <- numeric(0)
 
       for (i in seq_len(length(cpt.locs))) {
-        p.vals <- c(p.vals, sum(Tstar >= test.stat[cpt.locs[i]]) / reps)
+        scores <- c(scores, sum(test.stat[cpt.locs[i]] >= Tstar) / reps)
       }
     }
   } else {
     if (threshold == "bootstrap") {
-      p.vals <- sum(Tstar >= max(test.stat)) / reps
+      scores <- sum(max(test.stat) >= Tstar) / reps
     }
   }
 
@@ -276,7 +279,7 @@ np.mojo <- function(x, G, lag = 0, kernel.f = c("quad.exp", "gauss", "euclidean"
     use.mean = use.mean,
     test.stat = test.stat,
     cpts = cpt.locs,
-    p.vals = p.vals
+    scores = scores
   )
 
   return(ret)